+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Added build for linux appimage https://github.com/GrandOrgue/grandorgue/issues/698
 # 3.3.0 (2021-10-08)
 - Added automated release tagging on GitHub https://github.com/GrandOrgue/grandorgue/issues/711
 - Added polish language support https://github.com/GrandOrgue/grandorgue/discussions/743